pyproject.toml
[project]
name = "gemini-image-analyzer"
version = "0.1.0"
description = "Send images with prompts to Gemini API and get responses"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "python-docx>=1.0.0",
    "fpdf2>=2.7.0",
]

[project.scripts]
gemini-image = "gemini_image:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["gemini_image.py"]

